Update price integer -> bigInteger

+++ database/migrations/2026_05_15_125219_create_products_table.php
@@ -11,7 +11,7 @@ return new class extends Migration {
             $table->id();
             $table->string('name');
             $table->text('description');
-            $table->integer('price');
+            $table->bigInteger('price');
             $table->timestamps();
             $table->softDeletes();
         });
